refactor: swap optional exam 2 & 5
