+++ src/test/kotlin/ExamTest.kt
@@ -7,10 +7,10 @@ import com.dicoding.exam.latihan4.vehicle
 import com.dicoding.exam.latihan5.multiple
 import com.dicoding.exam.latihan5.sum
 import com.dicoding.exam.latihanopsional1.sumOfBigThree
-import com.dicoding.exam.latihanopsional2.concatString
+import com.dicoding.exam.latihanopsional2.minAndMax
 import com.dicoding.exam.latihanopsional3.manipulateString
 import com.dicoding.exam.latihanopsional4.getMiddleCharacters
-import com.dicoding.exam.latihanopsional5.minAndMax
+import com.dicoding.exam.latihanopsional5.concatString
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.fail
@@ -289,53 +289,30 @@ class ExamTest : Spek({
 
         describe("Pengecekan Latihan Opsional 2") {
             it("Latihan Opsional 2") {
-                try {
-                    Files.readAllLines(File("src/main/kotlin/com/dicoding/exam/latihanopsional2/App.kt").toPath())
-                        .filterIndexed { index, _ -> index >= 39 }
-                        .forEach { line ->
-                            assertEqualsWith(
-                                actual = "fun" in line,
-                                expected = false,
-                                errMessage = "concatString() tidak boleh menggunakan keyword `fun`"
-                            )
-                            assertEqualsWith(
-                                actual = "{" in line,
-                                expected = false,
-                                errMessage = "concatString() tidak boleh mengandung karakter `{`"
-                            )
-                            assertEqualsWith(
-                                actual = "}" in line,
-                                expected = false,
-                                errMessage = "concatString() tidak boleh menggunakan keyword `}`"
-                            )
-                        }
-                } catch (e: IOException) {
-                    fail("Failed to read the source file.")
-                }
-            }
-
-            it("Latihan Opsional 2") {
-                val result = concatString("Hello", "Dicoding")
+                val result = minAndMax(985684579)
                 assertEqualsWith(
                     actual = result,
-                    expected = "HelloDicoding",
-                    errMessage = "Jika argumennya adalah 'Hello' dan 'Dicoding', fungsi concatString() seharusnya mengembalikkan nilai HelloDicoding"
+                    expected = 13,
+                    errMessage = "Jika argumennya adalah '985684579', fungsi minAndMax() seharusnya mengembalikkan nilai '13'"
                 )
             }
 
             it("Latihan Opsional 2") {
-                val r = Random(System.currentTimeMillis())
-                repeat(3) {
-                    val string1 = r.nextInt().toString(2)
-                    val string2 = r.nextInt().toString(2)
-                    assertEqualsWith(
-                        actual = concatString(string1, string2),
-                        expected = string1 + string2,
-                        errMessage = "Jika argumennya adalah '$string1' dan '$string2', fungsi concatString() seharusnya mengembalikkan nilai ${
-                            string1.plus(string2)
-                        }"
-                    )
-                }
+                val result = minAndMax(1111111)
+                assertEqualsWith(
+                    actual = result,
+                    expected = 2,
+                    errMessage = "Jika argumennya adalah '1111111', fungsi minAndMax() seharusnya mengembalikkan nilai '2'"
+                )
+            }
+
+            it("Latihan Opsional 2") {
+                val result = minAndMax(6423654)
+                assertEqualsWith(
+                    actual = result,
+                    expected = 8,
+                    errMessage = "Jika argumennya adalah '6423654', fungsi minAndMax() seharusnya mengembalikkan nilai '8'"
+                )
             }
         }
 
@@ -400,30 +377,52 @@ class ExamTest : Spek({
 
         describe("Pengecekan Latihan Opsional 5") {
             it("Latihan Opsional 5") {
-                val result = minAndMax(985684579)
+                try {
+                    Files.readAllLines(File("src/main/kotlin/com/dicoding/exam/latihanopsional5/App.kt").toPath())
+                        .filterIndexed { index, _ -> index >= 39 }.forEach { line ->
+                            assertEqualsWith(
+                                actual = "fun" in line,
+                                expected = false,
+                                errMessage = "concatString() tidak boleh menggunakan keyword `fun`"
+                            )
+                            assertEqualsWith(
+                                actual = "{" in line,
+                                expected = false,
+                                errMessage = "concatString() tidak boleh mengandung karakter `{`"
+                            )
+                            assertEqualsWith(
+                                actual = "}" in line,
+                                expected = false,
+                                errMessage = "concatString() tidak boleh menggunakan keyword `}`"
+                            )
+                        }
+                } catch (e: IOException) {
+                    fail("Failed to read the source file.")
+                }
+            }
+
+            it("Latihan Opsional 5") {
+                val result = concatString("Hello", "Dicoding")
                 assertEqualsWith(
                     actual = result,
-                    expected = 13,
-                    errMessage = "Jika argumennya adalah '985684579', fungsi minAndMax() seharusnya mengembalikkan nilai '13'"
+                    expected = "HelloDicoding",
+                    errMessage = "Jika argumennya adalah 'Hello' dan 'Dicoding', fungsi concatString() seharusnya mengembalikkan nilai HelloDicoding"
                 )
             }
 
             it("Latihan Opsional 5") {
-                val result = minAndMax(1111111)
-                assertEqualsWith(
-                    actual = result,
-                    expected = 2,
-                    errMessage = "Jika argumennya adalah '1111111', fungsi minAndMax() seharusnya mengembalikkan nilai '2'"
-                )
-            }
-
-            it("Latihan Opsional 5") {
-                val result = minAndMax(6423654)
-                assertEqualsWith(
-                    actual = result,
-                    expected = 8,
-                    errMessage = "Jika argumennya adalah '6423654', fungsi minAndMax() seharusnya mengembalikkan nilai '8'"
-                )
+                val r = Random(System.currentTimeMillis())
+                repeat(3) {
+                    val string1 = r.nextInt().toString(2)
+                    val string2 = r.nextInt().toString(2)
+                    assertEqualsWith(
+                        actual = concatString(string1, string2),
+                        expected = string1 + string2,
+                        errMessage = "Jika argumennya adalah '$string1' dan '$string2', fungsi concatString() seharusnya mengembalikkan nilai ${
+                            string1.plus(string2)
+                        }"
+                    )
+                }
             }
         }
     }
